requirements.txt
pandas
pyarrow
python-dotenv
pandas-market-calendars
pytz
pyyaml
tqdm
numpy
